Feed a board into run_loop when called recursively.
The previous version had a bug in which it would feed the string version into the next iteration.  Actually, worse it would feed the result of the user provided function display into the next iteration.  I've moved the stringifcation and display into a function called output.  Life.output always returns the original board so that the pipeline in run_loop works as expected.

This required an update to the "It should run the game" test.  The test now inspects the output directly using CaptureIO.  I like this better than depending on the return value of display being propagated all the way out of run_loop.  I explored an alternative of asserting within the display function but this had the drawback that it would not fail the test if the display function was never called.

The multiple iterations test now passes and is no longer marked as pending.